feat: Refactor authentication context and update routing for role-based access control
