+++ client/src/contexts/AuthContext.jsx
